Move specimens to src/specimens
