+++ src/specimens/FramedCodeBlock.js
