+++ src/specimens/Hint.js
