+++ src/specimens/ReactComponent.js
