+++ src/specimens/Type.js
